requirement.txt
aiohttp==3.7.4.post0
altgraph==0.17
aniso8601==7.0.0
anyascii==0.1.7
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
bcrypt==3.2.0
beautifulsoup4==4.8.2
blinker==1.4
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
chromedriver-py==88.0.4324.27
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==3.4.6
distlib==0.3.1
Django==3.1.7
django-extensions==3.1.1
django-filter==2.4.0
django-modelcluster==5.1
django-rest-swagger==2.2.0
django-taggit==1.3.0
django-treebeard==4.4
djangorestframework==3.12.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
Flask-Breadcrumbs==0.5.1
Flask-Humanize==0.3.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Menu==0.7.2
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
fpdf==1.7.2
future==0.18.2
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
html5lib==1.1
humanize==3.2.0
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
Js2Py==0.70
jwt==1.2.0
keepa==1.2.1
l18n==2020.6.1
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
multidict==5.1.0
mysql-connector-python==8.0.22
mysqlclient==2.0.3
Naked==0.1.31
numpy==1.20.1
objsize==0.3.3
openapi-codec==1.3.2
openpyxl==3.0.4
pandas==1.2.3
paramiko==2.7.2
pathlib==1.0.1
pdfrw==0.4
pefile==2019.4.18
Pillow==8.0.1
progressbar==2.5
promise==2.3
protobuf==3.13.0
psycopg2==2.8.6
py3dns==3.2.1
PyAutoGUI==0.9.52
pycodestyle==2.7.0
pycparser==2.20
PyGetWindow==0.0.9
PyInstaller==3.6
pyjsparser==2.7.1
pylint==2.5.3
pylint-django==2.4.2
pylint-plugin-utils==0.6
pymongo==3.11.3
PyMsgBox==1.0.9
PyNaCl==1.4.0
pyotp==2.6.0
PyPDF2==1.26.0
pyperclip==1.8.1
PyQt5==5.15.1
PyQt5-sip==12.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pyserial==3.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-magic-bin==0.4.14
PyTweening==1.0.3
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
pyzmq==22.0.3
reportlab==3.5.54
requests==2.24.0
Rx==1.6.1
scraperapi-sdk==0.2.2
selenium==3.141.0
shellescape==3.8.1
simple-image-download==0.4
simplejson==3.17.2
singledispatch==3.6.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
ssh-pymongo==1.0.4
sshtunnel==0.4.0
stripe==2.56.0
tablib==3.0.0
telepot==12.7
text-unidecode==1.3
toml==0.10.1
tqdm==4.59.0
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
validate-email==1.3
virtualenv==20.4.2
wagtail==2.12.2
webencodings==0.5.1
Werkzeug==1.0.1
Willow==1.4
wrapt==1.12.1
WTForms==2.3.3
xlrd==1.2.0
XlsxWriter==1.3.6
xlwt==1.3.0
yarl==1.6.3
zmq==0.0.0
